refactor(rust): split commands/groups.rs into submodules (groups, channels, membership, invites, join_requests)
