+++ pollis-core/src/commands/groups/types.rs
@@ -1,0 +1,62 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Group {
+    pub id: String,
+    pub name: String,
+    pub description: Option<String>,
+    pub owner_id: String,
+    pub created_at: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Channel {
+    pub id: String,
+    pub group_id: String,
+    pub name: String,
+    pub description: Option<String>,
+    // 'text' or 'voice' — persisted in Turso.
+    // Migration: ALTER TABLE channels ADD COLUMN channel_type TEXT NOT NULL DEFAULT 'text';
+    pub channel_type: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct GroupWithChannels {
+    pub id: String,
+    pub name: String,
+    pub description: Option<String>,
+    pub owner_id: String,
+    pub created_at: String,
+    pub current_user_role: String,
+    pub channels: Vec<Channel>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct GroupMember {
+    pub user_id: String,
+    pub username: Option<String>,
+    pub display_name: Option<String>,
+    pub avatar_url: Option<String>,
+    pub role: String,
+    pub joined_at: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct PendingInvite {
+    pub id: String,
+    pub group_id: String,
+    pub group_name: String,
+    pub inviter_id: String,
+    pub inviter_username: Option<String>,
+    pub created_at: String,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct JoinRequest {
+    pub id: String,
+    pub group_id: String,
+    pub requester_id: String,
+    pub requester_username: Option<String>,
+    pub status: String,
+    pub created_at: String,
+}
